ref(backup): Re-organize test helpers
More of these helpers methods will be used by multiple test files, so
I've moved them into `__init__.py` in a separate commit before those
actual changes come in, for smaller diff sizes.
